# [Azure Service Bus] Remove unnecessary LINQ on AmqpReceiver