@pakrym 2020-04-13
Is there a benchmark/profile showing the improvement?