@pakrym 2020-04-13
> @pakrym happy to provide that if you think this is worth pursuing.

It's always good to have an idea of before/after and how the improvement compares to overall performance.