@danielmarbach 2020-04-13
@pakrym added the benchmark results with the changed code that hopefully also addresses the other comments. I ended up with a combination of LINQ and foreach to avoid the Any which leads to very good results.

I'm executing a medium run as I post this and will update the post as soon as I have the results

I'll leave it up to you if you want to take this or not. I have no problems closing this.