@AlexanderSher 2020-04-13
The best gain in the test results above is for the array of 4 elements and it is 175 ns. How many times this method is called?