@JoshLove-msft 2020-04-13
@danielmarbach thanks for these performance fixes! They are greatly appreciated 😄 Out of curiosity, have you profiled any end to end scenarios like receiving or sending to identify any bottlenecks? If so, we would be interested to learn about your findings.