@AlexanderSher 2020-04-13
I'm sorry for being stubborn, but I'm still interested in the scenario we're optimizing. I've done a simple benchmark based on the https://github.com/danielmarbach/MicroBenchmarks/commit/7a474da91836932c59a5cec74344e789453bd02e#diff-5d50ca96fc4df16892b002db1d70e8c3:

```cs
private static async Task<Guid[]> LinqWithAwait(IEnumerable<string> lockTokens)
{
    var lockTokenGuids = lockTokens.Select(token => new Guid(token)).ToArray();
    if (lockTokenGuids.Any(lockToken => _requestResponseLockedMessages.Contains(lockToken)))
        return await Task.FromResult(lockTokenGuids).ConfigureAwait(false);

    return await Task.FromResult(Array.Empty<Guid>()).ConfigureAwait(false);
}

private static async Task<Guid[]> ForeachWithAwait(IEnumerable<string> lockTokens)
{
    var requestResponse = false;
    var lockTokenGuids = lockTokens.Select(token => new Guid(token)).ToArray();
    foreach (var tokenGuid in lockTokenGuids)
    {
        if (!requestResponse && _requestResponseLockedMessages.Contains(tokenGuid)) requestResponse = true;
    }

    if (requestResponse)
    {
        return await Task.FromResult(lockTokenGuids).ConfigureAwait(false);
    }

    return await Task.FromResult(Array.Empty<Guid>()).ConfigureAwait(false);
}

private static Task<Guid[]> LinqWithoutAwait(IEnumerable<string> lockTokens)
{
    var lockTokenGuids = lockTokens.Select(token => new Guid(token)).ToArray();
    if (lockTokenGuids.Any(lockToken => _requestResponseLockedMessages.Contains(lockToken)))
        return Task.FromResult(lockTokenGuids);

    return Task.FromResult(Array.Empty<Guid>());
}
```

Here are the results:
|           Method |           Job |       Runtime | Elements |       Mean |    Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |-------------- |-------------- |--------- |-----------:|---------:|----------:|------:|--------:|-------:|------:|------:|----------:|
|    LinqWithAwait |      .NET 4.8 |      .NET 4.8 |        4 | 2,767.8 ns | 55.45 ns | 118.16 ns |  1.00 |    0.00 | 0.0687 |     - |     - |     449 B |
| ForeachWithAwait |      .NET 4.8 |      .NET 4.8 |        4 | 2,708.8 ns | 54.26 ns | 115.64 ns |  0.98 |    0.06 | 0.0648 |     - |     - |     417 B |
| LinqWithoutAwait |      .NET 4.8 |      .NET 4.8 |        4 | 2,608.3 ns | 52.26 ns | 147.39 ns |  0.95 |    0.07 | 0.0572 |     - |     - |     369 B |
|                  |               |               |          |            |          |           |       |         |        |       |       |           |
|    LinqWithAwait | .NET Core 2.1 | .NET Core 2.1 |        4 | 1,933.2 ns | 38.37 ns |  71.11 ns |  1.00 |    0.00 | 0.0648 |     - |     - |     416 B |
| ForeachWithAwait | .NET Core 2.1 | .NET Core 2.1 |        4 | 1,870.9 ns | 37.34 ns | 107.74 ns |  0.97 |    0.05 | 0.0572 |     - |     - |     384 B |
| LinqWithoutAwait | .NET Core 2.1 | .NET Core 2.1 |        4 | 1,997.0 ns | 54.25 ns | 159.97 ns |  0.98 |    0.06 | 0.0534 |     - |     - |     344 B |
|                  |               |               |          |            |          |           |       |         |        |       |       |           |
|    LinqWithAwait | .NET Core 3.1 | .NET Core 3.1 |        4 | 1,061.0 ns | 23.19 ns |  68.36 ns |  1.00 |    0.00 | 0.0515 |     - |     - |     416 B |
| ForeachWithAwait | .NET Core 3.1 | .NET Core 3.1 |        4 |   986.7 ns | 19.76 ns |  57.63 ns |  0.94 |    0.08 | 0.0477 |     - |     - |     384 B |
| LinqWithoutAwait | .NET Core 3.1 | .NET Core 3.1 |        4 |   974.0 ns | 19.89 ns |  55.46 ns |  0.92 |    0.06 | 0.0439 |     - |     - |     344 B |

As you can see, simple removal of `async/await` in private method provides better results than LINQ replaced with `foreach`, both in speed and memory allocation. Yet we are still talking about nanoseconds and bytes here. This code can be improved even further, with total removal of LINQ and casting to array so that `foreach` doesn't allocate any memory, but will anyone see these optimizations in the shadow of `DisposeMessageRequestResponseAsync`?